+++ jeap-spring-boot-vault-starter/src/main/java/ch/admin/bit/jeap/vault/config/JeapVaultKubernetesConfigLogger.java
@@ -4,8 +4,8 @@ import ch.admin.bit.jeap.vault.config.conditions.OnVaultAndKubernetesCondition;
 import jakarta.annotation.PostConstruct;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.context.annotation.Conditional;
-import org.springframework.context.annotation.Configuration;
 
 /**
  * Configuration class for logging Vault integration settings.
@@ -24,7 +24,7 @@ import org.springframework.context.annotation.Configuration;
  *
  * @see OnVaultAndKubernetesCondition
  */
-@Configuration
+@AutoConfiguration
 @Conditional(OnVaultAndKubernetesCondition.class)
 @Slf4j
 public class JeapVaultKubernetesConfigLogger {