+++ jeap-spring-boot-vault-starter/src/test/java/ch/admin/bit/jeap/vault/NoVaultIT.java
@@ -1,7 +1,5 @@
 package ch.admin.bit.jeap.vault;
 
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,8 +8,8 @@ import org.springframework.test.context.ActiveProfiles;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SpringBootTest
-@ActiveProfiles("local")
-public class NoVaultNoBootstrapIT {
+@ActiveProfiles("local") // note: no vault profile selected
+public class NoVaultIT {
 
     @Autowired
     private VaultTestConfiguration vaultTestConfiguration;
@@ -21,13 +19,4 @@ public class NoVaultNoBootstrapIT {
         assertEquals("local-secret-value", vaultTestConfiguration.getTestSecret());
     }
 
-    @BeforeAll
-    static void init()  {
-        System.setProperty("spring.cloud.bootstrap.enabled", "false");
-    }
-
-    @AfterAll
-    static void reset() {
-        System.clearProperty("spring.cloud.bootstrap.enabled");
-    }
 }