+++ jeap-spring-boot-vault-starter/src/test/java/ch/admin/bit/jeap/vault/VaultTestContainer.java
@@ -10,10 +10,5 @@ class VaultTestContainer extends VaultContainer<VaultTestContainer> {
 
     VaultTestContainer() {
         super(IMAGE_NAME);
-        // Vault configuration properties must be part of the bootstrap application context, as they are used to load
-        // properties to create the actual spring application context. Dynamically adding bootstrap properties
-        // (i.e. a random vault container port) in spring boot tests is extremely cumbersome. Thus using a fixed port
-        // here.
-        addFixedExposedPort(28282, 8200);
     }
 }