+++ jeap-spring-boot-vault-starter/src/test/java/ch/admin/bit/jeap/vault/VaultTestcontainersIT.java
@@ -1,17 +1,11 @@
 package ch.admin.bit.jeap.vault;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.context.ActiveProfiles;
 
-@ActiveProfiles("vault")
+@ActiveProfiles({"vault", "import"})
 @SuppressWarnings("java:S2699") // asserts are in the super class method, but sonar does not get it
-class VaultTestcontainersBootstrapContextIT extends VaultTestcontainersBaseIT {
-
-    @BeforeAll
-    static void init() {
-        init(true);
-    }
+class VaultTestcontainersIT extends VaultTestcontainersBaseIT {
 
     @Test
     void contextLoadsWithSecretsInjectedFromVaultInTestcontainer() {